pyproject.toml
[project]
name = "camera-db"
version = "0.1.0"
description = "Camera management system with FastAPI and SQLModel"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.1",
    "fastapi>=0.121.0",
    "sqlmodel>=0.0.27",
    "uvicorn>=0.38.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
]
